Create n_step_rnn
