+++ chainer/functions/__init__.py
@@ -49,6 +49,7 @@ from chainer.functions.connection import dilated_convolution_2d
 from chainer.functions.connection import embed_id
 from chainer.functions.connection import linear
 from chainer.functions.connection import n_step_lstm
+from chainer.functions.connection import n_step_rnn
 from chainer.functions.evaluation import accuracy
 from chainer.functions.evaluation import binary_accuracy
 from chainer.functions.evaluation import classification_summary \
@@ -206,6 +207,7 @@ deconvolution_nd = deconvolution_nd.deconvolution_nd
 dilated_convolution_2d = dilated_convolution_2d.dilated_convolution_2d
 embed_id = embed_id.embed_id
 linear = linear.linear
+NStepRNN = n_step_rnn.NStepRNN
 NStepLSTM = n_step_lstm.NStepLSTM
 n_step_lstm = n_step_lstm.n_step_lstm
